tallerApis/app/requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
PyJWT==2.8.0
Flask-JWT-Extended==4.3.0